Edit button refresh bugfix (#174)
* Add refreshControl to YAML and Violation tabs

Loading the Violation and YAML tab URLs directly was not loading the refresh module, which is dependent upon the refreshControl object being instantiated. This fix is to pass the refreshControl object to the Violation and YAML tabs from PolicyDetail.

* Allow Edit button to stay active after submit

After submitting, the edit button was responding to the refresh, making subsequent edits frustrating since it would automatically switch to read-only. This adjusts that logic.

However, the side effect is that the resourceData may not have yet caught up with the request, so clicking Submit again immediately after editing for a second time will result in a parsing error, but it seems the chances someone will click on the "Submit" button so quickly may be relatively low.

* Remove unnecessary filter update

The Filter portion of the toolbar isn't even showing, so it's unclear why code is needed to update the activeFilters in the state when it most likely hasn't changed.

* Allow polling for all but violations tab

Clicking the violations tab would stop polling for other tabs since the refresh state was being passed now that refresh controls were present. The violations tab has its own query interval based on the refresh, which is why polling is stopped for auto refresh but the controls are still relevant.

* Replace deprecated React methods

Changes:
- UNSAFE_componentWillReceiveProps is being deprecated, so this implements componentDidUpdate in its place
- Implement a check since calling startPolling with no pollInterval breaks the UI
- Remove the breadcrumb call from the ViolationTab since it's called by its parent and is unnecessary

* Fix coding bug

Co-authored-by: Yu Cao <ycao@redhat.com>