Summary page Filter E2E tests (#178)
* Fix filter functionality

Locally the filters were not displaying because our filters are saved as sets and sets cannot be stringified for the sessionStorage object. The fix converts these sets to arrays so that they can be stringified for the sessionStorage. Additionally, we're deleting the value before setting it when it appears that just setting it is sufficient.

* tested and polish

* Update filter-helper.js

fixed typo

* Summary page filters E2E test

Test filters on summary page, including whether the filters persist when navigating away from the page

* revert

Co-authored-by: ChunxiAlexLuo <55596988+ChunxiAlexLuo@users.noreply.github.com>
Co-authored-by: Yu Cao <ycao@redhat.com>