E2E RBAC Test Suite (#169)
* User setup and login/logout flows

- Added commands and YAML to create htpasswd users in the Hub cluster on build
- Added to gitignore in case someone creates an htpasswd file and it's not deleted
- Added initial tests to login and logout as the RBAC users

* Refactor RBAC build scripts; Use RBAC user for E2E

- Pulls out the RBAC portion into its own script, `build/rbac-setup.sh`, called from `build/run-e2e-tests.sh`
- Removes RBAC cleanup code from build (but script is still available in `rbac-clean.sh`)
- Adds encrypted RBAC_PASS variable to Travis environment variables
- Adds a `view` role in the `default` namespace to the namespaced admin user for testing

* Workaround for collectCoverage() error

Logging in and logging out seems to be causing errors with the coverage collection occuring at the end of the test (with or without instrumentation running).
This issue appears to match documented a documented known issue about page changes: https://github.com/aberonni/nightwatch-coverage#known-limitations

For reference, the error is:
```
{
  sessionId: 'e7e0fe7012747ae3d567402e7ac507b6',
  status: 0,
  value: null
}
   Error: Error while running "execute" command: An error occurred while gathering coverage information
       at runMicrotasks (<anonymous>)
       at processTicksAndRejections (internal/process/task_queues.js:97:5)
```

* Remove rbac-setup.sh script argument

Argument was unnecessary since the TRAVIS_BUILD_DIR would be available to the script already

* Test setup and initial summary test suite

- Create two namespaces for testing and don't use "default" since other tests may be using it
- Create policies before test and delete after
- Make sure each user sees the correct number of rows in the summary table based on namespace access

* Remove IBM licensing

* Add htpasswd install command to E2E build script

* Remove IBM licensing

* Move install from script to travis.yml

* Revert "Move install from script to travis.yml"

This reverts commit 8ac27da290c94dfe43adef0d3a3845c0b42042e4.

* Update command from apt to apt-get

* RBAC_PASS moved to Travis repo

Remove RBAC_PASS from .travis.yml in favor of keeping it in the Travis repo itself. (Deleting the e2e-users secret in openshift-config namespace should allow the password to update on the next build.)

* Implement DISABLE_CANARY_TEST variable

Replaces the DISABLE_CERT_TEST variable with a more generic one so that it can be used with RBAC (and any other future tests that may require something environment-specific)

* Update README

- Add RBAC information
- Remove no longer used information
- Update information to match current development flows and provide additional details

* Update README numbering

* Test suites for current UI behaviors

Tests RBAC access, but permissions are set for the current UI setup where all users see that everything is enabled

* Run E2E RBAC subscript in same shell

Exported variables are not currently being read by the parent E2E script since the script is being run as a subscript in its own shell

* E2E RBAC tests for create URL and edit permissions

- Test to make sure users without "create" permissions can't get to the create page through the URL
- Test to make sure the "edit" and "submit" buttons are disabled on the policy YAML tab for users without "patch" permissions

* Adjust tests for RBAC code

Originally the tests were set up and modified for the current UI behaviors. This removes these modifications.

* Fix "element not interactable" error

The DOM is reloading after the spinner disappears. This adds a 1 second wait to compensate.

* Add test for Placement edit buttons

Add test to check to make sure Placement Rule/Binding buttons are enabled/disabled

* Lengthen wait time for DOM reload

* Add wait for element for Placement edit check

* Fix action menu options

The Cluster Admin menu actions were not being updated because we were modifying the actions directly. This creates a clone so that the original remains unchanged.

* Fix E2E errors

Errors were occurring with DOM updates (perhaps due to the new ChromeDriver version). This implements additional waits to compensate.

Co-authored-by: Yu Cao <ycao@redhat.com>
Co-authored-by: Alex Luo <55596988+ChunxiAlexLuo@users.noreply.github.com>