Table search bar (#217)
* Insert search component above table

Only the component--doesn't filter table contents yet.

* Implement search filter function

* Update filtering behavior

The initial implementation was only filtering each page of results. This leverages the getDerivedStateFromProps where the state is being set with the rows in order to perform the correct filtering and deliver the rows to the render().

* Remove i18n dependency for placeholder

Since this is a reusable component, placeholder string will be passed in directly.

* Update unit tests

* Update unit tests

* Pagination togglea and styling

The pagination menu was not being styled with a white background
This also allows pagination to be disabled if desired

* Remove unnecessary CSS import

* Initial solution for non-string cells

The original filter function only dealt with cells containing strings. This is an initial implementation of a cell that contains HTML or React content.

* Update for various cell types in Patternfly

* Update styling

* Add padding for empty content icon

* Add fallback for filter logic

Co-authored-by: Yu Cao <ycao@redhat.com>