Use common ui-components (#605)
* Replace first PolicyStatusView with an AcmTable

The `transform_new` util function is also being worked on by @dhaiducek
in his branch. Some other pieces were taken from the old GRC
PatternFlyTable implementation.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use AcmTable in byTemplate tables in status view

Slowly removing the uses of the GRC implementation of PatternFlyTable.

As a result of how the searchbar-controlling function is implemented,
the search query from the "Clusters" tab is carried over into every
table under the "Templates" tab.

The tab-switcher between the Clusters and Templates view is repeated on
each table under the "Templates" tab, which might be wrong, but moving
it outside of the table would make it different than how it is on the
"Clusters" tab, which seems to be the 'correct' way to do it...

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Refactor groupByTemplate

Previous implementation was highly imperative, and imo hard to follow.
This implementation makes more use of "functional" javascript features,
but it might be less performant since it goes over the data more times.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use AcmTable in history view page

The default search placeholder is now 'Search', but it was 'Find'
before.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use AcmTable in the template details view

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use AcmTable in Policy Details Overview

The AcmTable doesn't (currently) allow removal of the search bar, but
the pagination is able to be automatically hidden.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use AcmDescriptionList

The extra logic around formatting the list items in the overview page
was taken (with slight updates) from the DetailsModule. Some logic
regarding tooltips was removed because I couldn't find any instances
where it was actually used.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Remove DetailsModule

The one place this was used has been updated to use AcmDescriptionList

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Move cluter/template toggle out of table

This returns the toggle to its previous position, it doesn't make sense
to have it in the tables because the Templates page can have multiple
tables.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* CSS Fixes

Margin and padding adjustments. In older versions of the ui-components,
there were additional `div`s under the `page-content-container`, but
that was changed in ui-component PR 310 / v0.78.0.

This should preserve the original look and feel.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Fix shared search values in Templates view

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Update transform_new function

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Fix sorting

The AcmTable actually takes the sortBy setting just as `sort`, but then
we would need to supply a corresponding `setSort` function for it to
work. Since these tables don't have a default `sortBy` anyway, let's
just take the default behavior.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Fix cypress tests

Some of the details (class names, ids, structure) were changed between
the PatternFlyTable which we previously tested against, and the new
AcmTable.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Update snapshots

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>