Sort tables (#620)
* Add initial sorting to AcmTables

With the most recent release of ui-components, it is now easy to set an
initial sorting on the table, and not have to manage the sortBy state
directly.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12610

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Update snapshots

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Fix test for Create page

With an update to the AcmHeader in ui-components, Openshift console
pages (including the "Create" page) will open in a new window. See
https://github.com/open-cluster-management/ui-components/pull/344

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>