Timestamp sorting (#633)
* Use type: timestamp instead of transformFunction

As long as `resourceKey` is set correctly, they would accomplish the
same thing. I think it is helpful for everything to go through the
timestamp creation logic in transformNew, since the `rawData` that it
adds will be useful.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Use sortable: true instead of transform

AcmTable will automatically add the sortable transform to the table
items if we give it a sort property. Using a sort property will let it
sort on things that aren't exactly what is in the cell, which is
necessary if the cell has a link or something in it.

If both AcmTable and our TableDefs add the sortable transform, then the
table header will have two sortable arrows, which is not good.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12761

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Fix sorting by using sortLabels

The sortLabel can be used on table items that don't just contain text,
for example if they contain a link or timestamp, this is necessary. In
most of these cases, the `rawData` is supplied by a transformFunction.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12761

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Remove old transform function

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Adjust imagemanifest test policy names

The AcmTable fuzzy searching makes some tests more fragile. If two
policies have long names that are relatively similar, then searches
often return both, which can lead to test failures. This updates some
of the test policy names to make them shorter and more unique.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Update snapshots

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>

* Suggestions from review

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>